Write long size instead of int, and convert when allocating byte array

+++ chill-hadoop/src/main/java/com/twitter/chill/hadoop/KryoSerializer.java
@@ -47,7 +47,7 @@ public class KryoSerializer implements Serializer<Object> {
         try {
           st.writeObject(o);
           // Copy from buffer to output stream.
-          outputStream.writeInt(st.numOfWrittenBytes());
+          outputStream.writeLong(st.numOfWrittenBytes());
           st.writeOutputTo(outputStream);
         }
         finally {
